refactor(deepseek): 隔离 Legacy Host Adapter
