+++ packages/adapters/deepseek-harness/src/index.ts
@@ -1,12 +1,12 @@
 import { packageMetadata as harnessAdapter } from "@codexhost/harness-adapter";
 import { WORKSPACE_CONTRACT_VERSION } from "@codexhost/shared-contracts";
 
-export { DeepSeekHarnessAdapter } from "./deepseek-harness-adapter.js";
+export { DeepSeekHarnessAdapter } from "./legacy/deepseek-harness-adapter.js";
 export type {
   DeepSeekHarnessAdapterDependencies,
   DeepSeekHarnessAdapterOptions,
   DeepSeekHostConnectionLike,
-} from "./deepseek-harness-adapter.js";
+} from "./legacy/deepseek-harness-adapter.js";
 
 export const packageMetadata = {
   name: "@codexhost/adapter-deepseek-harness",
