+++ packages/adapters/deepseek-harness/src/legacy/host-client.ts
