+++ packages/adapters/deepseek-harness/src/legacy/permission-modes.ts
